Respond with content of requested file
When responding to requests we send a debug response, which is just a copy
of the request that the user agent sent. While that is useful we really
want to send the file that the user agent requested.

Look up the requested file and send it back to the user agent if it is
in the list of shared files.

Signed-off-by: Marcus Karlsson <mk@acc.umu.se>