Don't allow adding files with duplicate file name
We have not implemented support for subdirectories, which means that we
can't share multiple files with the same file name. We don't check for
this when new files are added. If the file name is requested then the
first one found will be returned.

Check for duplicate file names and prevent them from being added when
adding new files.

Signed-off-by: Marcus Karlsson <mk@acc.umu.se>