Reset port text box color when parsing succeeds
When we successfully parse the port number we remove the background. This
will not reset the background to the default, it will make it transparent.
This make the control look like it has the same color as the window, which
will sometimes make the control hard to read.

Clear the value of the background property when parsing succeeds. Doing
that will give it the current default color which is usually a good
choice.

Signed-off-by: Marcus Karlsson <mk@acc.umu.se>